Move SVG geometry patch to MAIN world via manifest
MV3 content scripts run in an isolated JS world by default. That means
`Element.prototype.setAttribute` patches in content.js only affect the
isolated world's Element.prototype — NOT the page's. Page scripts (which
drive Fidelity's pie chart animation via setAttribute('d', ...)) see the
unpatched original function, so the fix had no effect in real Chrome
deployment.

The previous Playwright test (which injects content.js via addScriptTag,
running in the page's main world) happened to pass because addScriptTag
doesn't get the isolated-world treatment — but that meant the test was
giving a false positive for real extension behavior. Confirmed by real-
user manual retest: pie chart still animated after the previous commit.

Fix: split the prototype patch into its own file, main-world-patch.js,
and add a second content_scripts entry in manifest.json with
"world": "MAIN". That runs the patch at document_start in the page's
actual JS world so page scripts see the patched setAttribute.

content.js still ships the CSS rule `svg [data-still-svg-settling] {
visibility: hidden !important }` via its <style> injection — CSS works
cross-world because it applies to rendered DOM, not world-specific
prototypes.

Updated test to inject main-world-patch.js before content.js, matching
the new load order defined in manifest.json.

26/26 tests passing. Real-extension retest to verify live Fidelity.
