+++ Still/Still/Still Extension/Resources/manifest.json
@@ -33,6 +33,12 @@
   "content_scripts": [
     {
       "matches": ["<all_urls>"],
+      "js": ["main-world-patch.js"],
+      "world": "MAIN",
+      "run_at": "document_start"
+    },
+    {
+      "matches": ["<all_urls>"],
       "js": ["content.js"],
       "run_at": "document_start"
     }
